feat(squid-proposals): moved file to correct directory
