+++ pkg/squid/src/@types/pallets/signal/proposalSimpleVoteData.ts
